+++ packages/credential-sdk/src/types/signatures/index.js
@@ -1,10 +1,12 @@
 import SignatureSr25519Value from './signature-sr25519-value';
 import SignatureEd25519Value from './signature-ed25519-value';
 import SignatureSecp256k1Value from './signature-secp256k1-value';
+import SignatureSecp256r1Value from './signature-secp256r1-value';
 
 export {
   SignatureSr25519Value,
   SignatureEd25519Value,
   SignatureSecp256k1Value,
+  SignatureSecp256r1Value,
 };
 export * from './signature';