+++ packages/credential-sdk/src/vc/crypto/EcdsaSecp256r1VerificationKey2019.js
@@ -1,0 +1,58 @@
+import b58 from 'bs58';
+import * as base64 from '@juanelas/base64';
+import { u8aToU8a } from '../../utils/types/bytes';
+import { EcdsaSecp256r1VerKeyName } from './constants';
+import Secp256r1Keypair from '../../keypairs/keypair-secp256r1';
+import { decodeFromMultibase } from '../../utils/encoding/multibase';
+
+export default class EcdsaSecp256r1VerificationKey2019 {
+  constructor(publicKey) {
+    this.publicKey = u8aToU8a(publicKey);
+  }
+
+  static from(verificationMethod) {
+    if (
+      !verificationMethod.type
+      || verificationMethod.type.indexOf(EcdsaSecp256r1VerKeyName) === -1
+    ) {
+      throw new Error(
+        `verification method should have type ${EcdsaSecp256r1VerKeyName} - got: ${verificationMethod.type}`,
+      );
+    }
+
+    if (verificationMethod.publicKeyBase58) {
+      return new this(b58.decode(verificationMethod.publicKeyBase58));
+    }
+
+    if (verificationMethod.publicKeyBase64) {
+      return new this(base64.decode(verificationMethod.publicKeyBase64));
+    }
+
+    const secMultibase = verificationMethod['sec:publicKeyMultibase'];
+    const multiBase = verificationMethod.publicKeyMultibase || (
+      typeof secMultibase === 'object'
+        ? secMultibase['@value']
+        : secMultibase
+    );
+
+    if (multiBase) {
+      return new this(decodeFromMultibase(multiBase).slice(2));
+    }
+
+    throw new Error(
+      `Unsupported signature encoding for ${EcdsaSecp256r1VerKeyName}`,
+    );
+  }
+
+  verifier() {
+    return EcdsaSecp256r1VerificationKey2019.verifierFactory(this.publicKey);
+  }
+
+  static verifierFactory(publicKey) {
+    return {
+      async verify({ data, signature }) {
+        return Secp256r1Keypair.verify(data, signature, publicKey);
+      },
+    };
+  }
+}